refac: FDD foldering
